refactor: move `MutableMessagesContainer` to its own file and apply the delegate pattern.
